freeze migrations into v2.4.0_2.ts, bump to 2.4.0:3
